Extract portable logic into @qali/domain package
Stand up a source-only @qali/domain workspace package (mirroring @qali/env)
and move the three browser-safe leaf modules — availability, permissions, slug
— plus their unit tests out of convex/lib into it. These import nothing from
Convex, so the frontend no longer reaches into backend deployment internals:
apps/web and the backend both consume @qali/domain/* instead of
@qali/backend/convex/lib/*.

Pure relocation — no behavior change. Git detects all five files as renames.
Verified: domain 46 tests + types clean; backend 70 unit + 18 integration +
types clean (convex/vite bundler resolves the workspace .ts exports); apps/web
builds and typechecks.
