+++ apps/web/package.json
@@ -18,6 +18,7 @@
     "@hugeicons/core-free-icons": "^4.2.2",
     "@hugeicons/react": "^1.1.9",
     "@qali/backend": "workspace:*",
+    "@qali/domain": "workspace:*",
     "@qali/env": "workspace:*",
     "@qali/ui": "workspace:*",
     "@tailwindcss/vite": "^4.3.2",
