+++ packages/domain/src/availability.test.ts
