+++ packages/domain/src/availability.ts
